bug(data_range): Replace float(data_range). Applie suggestions from code review

+++ piq/ssim.py
@@ -54,8 +54,8 @@ def ssim(x: torch.Tensor, y: torch.Tensor, kernel_size: int = 11, kernel_sigma: 
     x = x.type(torch.float32)
     y = y.type(torch.float32)
         
-    x = x / float(data_range)
-    y = y / float(data_range)
+    x = x / data_range
+    y = y / data_range
 
     kernel = gaussian_filter(kernel_size, kernel_sigma).repeat(x.size(1), 1, 1, 1).to(y)
     _compute_ssim_per_channel = _ssim_per_channel_complex if x.dim() == 5 else _ssim_per_channel
@@ -217,8 +217,11 @@ def multi_scale_ssim(x: torch.Tensor, y: torch.Tensor, kernel_size: int = 11, ke
     )
     x, y = _adjust_dimensions(input_tensors=(x, y))
 
-    x = x / float(data_range)
-    y = y / float(data_range)
+    x = x.type(torch.float32)
+    y = y.type(torch.float32)
+
+    x = x / data_range
+    y = y / data_range
 
     if scale_weights is None:
         scale_weights_from_ms_ssim_paper = [0.0448, 0.2856, 0.3001, 0.2363, 0.1333]
